Change vector type to long double

+++ Coding Challenges 2/SumFibonacciNumbers/fibonacci_utils.h
@@ -5,7 +5,7 @@
 
 namespace fibonacci
 {
-    std::vector<int> generate(int count);
+    std::vector<long double> generate(int count);
 }
 
 #endif  // !FIBONACCI_UTILS_H
